trailing colon question mark
Sometimes, the files of the ISO when mounted with fuseiso have a trailing
";?". Depending on the way fuseiso mounted the ISO (-c <charset>) option the
question mark '?' may be different, especially when mounted on with utf-8.

Better detected and manage this case.

See https://bugs.debian.org/867388

Gbp-Pq: Name 0017-trailing-colon-question-mark.patch