Improve `progressUpdate()` logic.
Only print a progress bar if the terminal is wide enough to fit it, otherwise
just print the raw percentage and ETA numbers.
Also, account for the fact that the '|' is printed even if the current progress
is zero, meaning we should always subtract at least one from the number of
space characters following it.